Adding a Picture of the first steps output
Before I actually start comparing the entropy of all EOWL hashes, I want to time the actual speed of the hashing itself. 
This picture shows three trials, where surprisingly the python hex encoding is the slowest process and the SHA-256 encryption is the fastest. The Hex time is not really all that slow, considering the size of the data but it's interesting.